Finallized implementation of EQ reports

+++ selenium-wd/tests/surveyor/scommon/source/Coordinates.java
@@ -2,35 +2,35 @@ package surveyor.scommon.source;
 
 public class Coordinates {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
-	public Coordinates (int x, int y){
+	public Coordinates (double x, double y){
 		this.x = x;
 		this.y = y;
 	}
 	/**
 	 * @return the x
 	 */
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	/**
 	 * @param x the x to set
 	 */
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	/**
 	 * @return the y
 	 */
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 	/**
 	 * @param y the y to set
 	 */
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 }
